move Vector class into core and spark.util package
